Moved out reader from native index accessor class
