+++ community/kernel/src/main/java/org/neo4j/kernel/impl/index/schema/NativeSchemaNumberIndexAccessor.java
@@ -22,26 +22,19 @@ package org.neo4j.kernel.impl.index.schema;
 import java.io.File;
 import java.io.IOException;
 import java.io.UncheckedIOException;
-import org.neo4j.collection.primitive.PrimitiveLongIterator;
-import org.neo4j.cursor.RawCursor;
+
 import org.neo4j.graphdb.ResourceIterator;
 import org.neo4j.helpers.collection.BoundedIterable;
-import org.neo4j.index.internal.gbptree.Hit;
 import org.neo4j.index.internal.gbptree.Layout;
 import org.neo4j.index.internal.gbptree.RecoveryCleanupWorkCollector;
 import org.neo4j.io.pagecache.IOLimiter;
 import org.neo4j.io.pagecache.PageCache;
 import org.neo4j.kernel.api.exceptions.index.IndexEntryConflictException;
-import org.neo4j.kernel.api.exceptions.index.IndexNotApplicableKernelException;
 import org.neo4j.kernel.api.index.IndexAccessor;
 import org.neo4j.kernel.api.index.IndexUpdater;
 import org.neo4j.kernel.api.index.PropertyAccessor;
-import org.neo4j.kernel.api.schema.IndexQuery;
-import org.neo4j.kernel.api.schema.IndexQuery.ExactPredicate;
-import org.neo4j.kernel.api.schema.IndexQuery.NumberRangePredicate;
 import org.neo4j.kernel.impl.api.index.IndexUpdateMode;
 import org.neo4j.storageengine.api.schema.IndexReader;
-import org.neo4j.storageengine.api.schema.IndexSampler;
 
 public class NativeSchemaNumberIndexAccessor<KEY extends NumberKey, VALUE extends NumberValue>
         extends NativeSchemaNumberIndex<KEY,VALUE> implements IndexAccessor
@@ -91,7 +84,7 @@ public class NativeSchemaNumberIndexAccessor<KEY extends NumberKey, VALUE extend
     @Override
     public IndexReader newReader()
     {
-        return new NativeSchemaNumberIndexReader();
+        return new NativeSchemaNumberIndexReader<>( tree, layout );
     }
 
     @Override
@@ -111,112 +104,5 @@ public class NativeSchemaNumberIndexAccessor<KEY extends NumberKey, VALUE extend
             throws IndexEntryConflictException, IOException
     {
         throw new UnsupportedOperationException( "Implement me" );
-    }
-
-    private class NativeSchemaNumberIndexReader implements IndexReader
-    {
-        private final KEY treeKeyFrom = layout.newKey();
-        private final KEY treeKeyTo = layout.newKey();
-        private RawCursor<Hit<KEY,VALUE>,IOException> openSeeker;
-
-        @Override
-        public void close()
-        {
-            ensureOpenSeekerClosed();
-        }
-
-        @Override
-        public long countIndexedNodes( long nodeId, Object... propertyValues )
-        {
-            treeKeyFrom.from( nodeId, propertyValues );
-            treeKeyTo.from( nodeId, propertyValues );
-            try ( RawCursor<Hit<KEY,VALUE>,IOException> seeker = tree.seek( treeKeyFrom, treeKeyTo ) )
-            {
-                long count = 0;
-                while ( seeker.next() )
-                {
-                    if ( seeker.get().key().entityId == nodeId )
-                    {
-                        count++;
-                    }
-                }
-                return count;
-            }
-            catch ( IOException e )
-            {
-                throw new UncheckedIOException( e );
-            }
-        }
-
-        @Override
-        public IndexSampler createSampler()
-        {
-            throw new UnsupportedOperationException( "Implement me" );
-        }
-
-        @Override
-        public PrimitiveLongIterator query( IndexQuery... predicates ) throws IndexNotApplicableKernelException
-        {
-            if ( predicates.length != 1 )
-            {
-                throw new UnsupportedOperationException();
-            }
-
-            ensureOpenSeekerClosed();
-            IndexQuery predicate = predicates[0];
-            switch ( predicate.type() )
-            {
-            case exists:
-                treeKeyFrom.initAsLowest();
-                treeKeyTo.initAsHighest();
-                return startSeekForInitializedRange();
-            case exact:
-                ExactPredicate exactPredicate = (ExactPredicate) predicate;
-                Object[] values = new Object[] {exactPredicate.value()};
-                treeKeyFrom.from( Long.MIN_VALUE, values );
-                treeKeyTo.from( Long.MAX_VALUE, values );
-                return startSeekForInitializedRange();
-            case rangeNumeric:
-                NumberRangePredicate rangePredicate = (NumberRangePredicate) predicate;
-                treeKeyFrom.from( rangePredicate.fromInclusive() ? Long.MIN_VALUE : Long.MAX_VALUE,
-                        new Object[] {rangePredicate.from()} );
-                treeKeyFrom.entityIdIsSpecialTieBreaker = true;
-                treeKeyTo.from( rangePredicate.toInclusive() ? Long.MAX_VALUE : Long.MIN_VALUE,
-                        new Object[] {rangePredicate.to()} );
-                treeKeyTo.entityIdIsSpecialTieBreaker = true;
-                return startSeekForInitializedRange();
-            default:
-                throw new IllegalArgumentException( "IndexQuery of type " + predicate.type() + " is not supported." );
-            }
-        }
-
-        private PrimitiveLongIterator startSeekForInitializedRange()
-        {
-            try
-            {
-                openSeeker = tree.seek( treeKeyFrom, treeKeyTo );
-                return new NumberHitIterator<>( openSeeker );
-            }
-            catch ( IOException e )
-            {
-                throw new UncheckedIOException( e );
-            }
-        }
-
-        private void ensureOpenSeekerClosed()
-        {
-            if ( openSeeker != null )
-            {
-                try
-                {
-                    openSeeker.close();
-                }
-                catch ( IOException e )
-                {
-                    throw new UncheckedIOException( e );
-                }
-                openSeeker = null;
-            }
-        }
     }
 }
